+++ packages/admin-panel/src/library.js
@@ -5,5 +5,5 @@
 import 'react-table/react-table.css';
 
 export * from './pages/resources';
-export * from './rootReducer';
+export * from './store';
 export * from './api';
