+++ packages/admin-panel/src/store.js
@@ -3,13 +3,17 @@
  *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 
+import React from 'react';
+import PropTypes from 'prop-types';
 import { createStore, applyMiddleware, compose } from 'redux';
+import { PersistGate } from 'redux-persist/lib/integration/react';
+import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import localforage from 'localforage';
 import { persistReducer, persistStore } from 'redux-persist';
-import { api } from './api';
 import { rootReducer } from './rootReducer';
 import { RememberMeTransform } from './authentication/reducer';
+import { ProfileButton } from './authentication';
 
 const persistedRootReducer = persistReducer(
   {
@@ -23,7 +27,6 @@ const persistedRootReducer = persistReducer(
 
 const initialState = {};
 const enhancers = [];
-const middleware = [thunk.withExtraArgument({ api })];
 
 if (process.env.NODE_ENV === 'development') {
   const { devToolsExtension } = window;
@@ -32,10 +35,31 @@ if (process.env.NODE_ENV === 'development') {
   }
 }
 
-const composedEnhancers = compose(applyMiddleware(...middleware), ...enhancers);
+export const StoreProvider = React.memo(({ children, api, persist }) => {
+  const middleware = [thunk.withExtraArgument({ api })];
+  const composedEnhancers = compose(applyMiddleware(...middleware), ...enhancers);
+  const store = createStore(persistedRootReducer, initialState, composedEnhancers);
+  const persistor = persistStore(store);
 
-export const store = createStore(persistedRootReducer, initialState, composedEnhancers);
+  if (!persist) {
+    return <Provider store={store}>{children}</Provider>;
+  }
 
-api.injectReduxStore(store);
+  return (
+    <Provider store={store}>
+      <PersistGate persistor={persistor} loading={null}>
+        {children}
+      </PersistGate>
+    </Provider>
+  );
+});
 
-export const persistor = persistStore(store);
+StoreProvider.propTypes = {
+  api: PropTypes.object.isRequired,
+  children: PropTypes.node.isRequired,
+  persist: PropTypes.bool,
+};
+
+StoreProvider.defaultProps = {
+  persist: false,
+};
