+++ packages/admin-panel/src/utilities/AdminPanelProviders.js
@@ -4,26 +4,23 @@
  */
 
 import React from 'react';
-import { Provider } from 'react-redux';
-import { PersistGate } from 'redux-persist/lib/integration/react';
 import { MuiThemeProvider, StylesProvider } from '@material-ui/core/styles';
 import { ThemeProvider } from 'styled-components';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import { persistor, store } from '../store';
+import { StoreProvider } from '../store';
 import { theme } from '../theme';
+import { api } from '../api';
 
 // eslint-disable-next-line react/prop-types
 export const AdminPanelProviders = ({ children }) => (
-  <Provider store={store}>
-    <PersistGate persistor={persistor} loading={null}>
-      <StylesProvider injectFirst>
-        <MuiThemeProvider theme={theme}>
-          <ThemeProvider theme={theme}>
-            <CssBaseline />
-            {children}
-          </ThemeProvider>
-        </MuiThemeProvider>
-      </StylesProvider>
-    </PersistGate>
-  </Provider>
+  <StoreProvider api={api} persist>
+    <StylesProvider injectFirst>
+      <MuiThemeProvider theme={theme}>
+        <ThemeProvider theme={theme}>
+          <CssBaseline />
+          {children}
+        </ThemeProvider>
+      </MuiThemeProvider>
+    </StylesProvider>
+  </StoreProvider>
 );
