--- packages/lesmis/src/admin-panel/apiSingleton.js
@@ -1,7 +1,0 @@
-/*
- * Tupaia
- *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-import { TupaiaApi } from './TupaiaApi';
-
-export const api = new TupaiaApi();
