--- packages/lesmis/src/admin-panel/index.js
@@ -1,5 +1,0 @@
-/*
- * Tupaia
- *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
- */
-export * from './store';
